@jataylo 2025-04-09
This failure is currently on the hud along side other benchmark failures
https://hud.pytorch.org/pytorch/pytorch/commit/604467de208646f0c3b2663e45f2ff6a655a6716

@eellison are we okay if I force push this?