@eellison 2025-04-09
@jataylo, yes, per the dr ci comment you should be good to land: https://github.com/pytorch/pytorch/pull/147315#issuecomment-2662636851